Add files via upload
Task 3 : IRIS FLOWER CLASSIFICATION
Objective is to train a machine learning model that can learn from
these measurements and accurately classify the Iris flowers into
their respective species.The task has to be done by using the DecisionTreeClassifier
A decision tree is a machine learning algorithm that learns a flowchart-like structure to make predictions. It splits the data based on features and creates decision rules. Each internal node represents a feature, each branch represents a decision, and each leaf node represents a prediction. It's used for classification and regression tasks and is known for its interpretability. Decision trees can handle different types of data, but they may overfit complex data.The code loads the Iris dataset, splits it into training and testing sets, trains a Decision Tree Classifier on the training data, predicts the target variable for the test data, and finally evaluates the model's accuracy in classifying the Iris flowers.
**Got accuracy 93.333% on testing data and have made prediction and got accurate results**